% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{get_piped_call_var}
\alias{get_piped_call_var}
\title{internalization of approach (C) Carl A. B. Pearson from
\href{https://github.com/pearsonca/cabputils}{cabputils}}
\usage{
get_piped_call_var(sc)
}
\description{
internalization of approach (C) Carl A. B. Pearson from
\href{https://github.com/pearsonca/cabputils}{cabputils}
}
